# swagger.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Swagger Integrity-First - OpenAPI 3.0",
    "description": "This is a Intergity-First Server based on the OpenAPI 3.0 specification.",
    "contact": {
      "email": "taha.muhammad.official@gmail.com"
    },
    "version": "1.0.11"
  },
  "servers": [
    {
      "url": "http://localhost:3002"
    },
    {
      "url": "https://staging.integrity1auto.com"
    },
    {
      "url": "https://api.integrity1auto.com"
    }
  ],
...
  "paths": {
...
    "/web-hooks/notification": {
      "post": {
        "tags": ["Webhooks"],
        "summary": "Save notification in DB",
        "security": [
          {
            "WebhookAPIKeyAuth": []
          }
        ],
        "requestBody": {
          "description": "",
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "title": {
                    "type": "string"
                  },
                  "description": {
                    "type": "string"
                  },
                  "kind": {
                    "type": "string"
                  },
                  "payload": {
                    "type": "object"
                  },
                  "userId": {
                    "type": "string"
                  }
                }
              }
            }
          },
          "required": true
        },
        "description": "",
        "responses": {
          "200": {
            "description": "Successful operation"
          }
        }
      }
    },
    "/web-hooks/redeem": {
      "post": {
        "tags": ["Webhooks"],
        "summary": "Redeem points",
        "security": [
          {
            "WebhookAPIKeyAuth": []
          }
        ],
        "requestBody": {
          "description": "",
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "points": {
                    "type": "number",
                    "example": 10
                  },
                  "userTekmetricId": {
                    "type": "number",
                    "example": 27800881
                  }
                }
              }
            }
          },
          "required": true
        },
        "description": "",
        "responses": {
          "200": {
            "description": "Successful operation"
          }
        }
      }
    },
    "/web-hooks/membership": {
      "post": {
        "tags": ["Webhooks"],
        "summary": "Redeem points",
        "security": [
          {
            "WebhookAPIKeyAuth": []
          }
        ],
        "requestBody": {
          "description": "",
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "startDate": {
                    "type": "string",
                    "example": "2023-01-24"
                  },
                  "vin": {
                    "type": "string",
                    "example": "3GYFNBE34ES560138"
                  },
                  "userTekmetricId": {
                    "type": "number",
                    "example": 27800881
                  }
                }
              }
            }
          },
          "required": true
        },
        "description": "",
        "responses": {
          "200": {
            "description": "Successful operation"
          }
        }
      }
    },
    "/web-hooks/points": {
      "post": {
        "tags": ["Webhooks"],
        "summary": "Reward points to user",
        "security": [
          {
            "WebhookAPIKeyAuth": []
          }
        ],
        "requestBody": {
          "description": "",
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "kind": {
                    "type": "string",
                    "example": "LOYALTY | REFERRAL"
                  },
                  "userTekmetricId": {
                    "type": "number",
                    "example": 27800881
                  }
                }
              }
            }
          },
          "required": true
        },
        "description": "",
        "responses": {
          "200": {
            "description": "Successful operation"
          }
        }
      }
    },
...
  },
  "components": {
...
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      },
      "WebhookAPIKeyAuth": {
        "type": "apiKey",
        "name": "Authorization",
        "in": "header"
      }
    }
  }
}